pyproject.toml
[tool.poetry]
name = "solenix test"
version = "0.1.0"
description = ""
authors = ["wassafshahzad <wassafshahzad@gmail.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.0.6"
djangorestframework = "^3.15.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
